requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
pillow==5.4.1
pytesseract==0.2.6
rpi.gpio==0.6.5
smbus2==0.2.3
